requirements.txt
dependency-injector>=4.41.0
pyodbc>=4.0.34
pyyaml>=6.0.0
questionary>=1.10.0
MetaTrader5>=5.0.44
pandas>=1.5.0
numpy>=1.23.0
matplotlib~=3.10.1
seaborn~=0.13.2
scikit-learn~=1.6.1
SQLAlchemy~=2.0.40